requirements.txt
https://github.com/Sandeepkandula2004/Cerevyn_22341A4217/releases/download/v1/insightface-0.7.3-cp312-cp312-linux_x86_64.whl
fastapi==0.124.4
uvicorn==0.38.0
pydantic==2.12.5
sqlalchemy==2.0.45
opencv-python==4.12.0.88
imgaug==0.4.0
faiss-cpu==1.13.1
supabase==2.25.1
python-dotenv==1.2.1
onnxruntime==1.23.2
psycopg2-binary==2.9.11
python-multipart==0.0.20
numpy==1.26.4